+++ R/validation-helpers.R
@@ -21,6 +21,7 @@
 #' validateIsString("x")
 #' validateIsLogical(TRUE)
 #' @export
+
 validateIsOfType <- function(object, type, nullAllowed = FALSE) {
   type <- c(type)
 
@@ -35,10 +36,11 @@ validateIsOfType <- function(object, type, nullAllowed = FALSE) {
 
   # Name of the variable in the calling function
   objectName <- deparse(substitute(object))
-  objectTypes <- typeNamesFrom(type)
+  objectTypes <- .typeNamesFrom(type)
 
   # There might be no call stack available if called from terminal
   callStack <- as.character(sys.call(-1)[[1]])
+
   # Object name is one frame further for functions such as ValidateIsNumeric
   if ((length(callStack) > 0) && grepl(pattern = "validateIs", x = callStack)) {
     objectName <- deparse(substitute(object, sys.frame(-1)))
@@ -57,6 +59,7 @@ validateIsString <- function(object, nullAllowed = FALSE) {
 
 #' @rdname validateIsOfType
 #' @export
+
 validateIsCharacter <- validateIsString
 
 #' @rdname validateIsOfType
@@ -134,9 +137,12 @@ validateIsLogical <- function(object, nullAllowed = FALSE) {
 
 validatePathIsAbsolute <- function(path) {
   wildcardChar <- "*"
-  if (any(unlist(strsplit(path, ""), use.names = FALSE) == wildcardChar)) {
-    stop(messages$errorEntityPathNotAbsolute(path))
+
+  if (!any(unlist(strsplit(path, ""), use.names = FALSE) == wildcardChar)) {
+    return()
   }
+
+  stop(messages$errorEntityPathNotAbsolute(path))
 }
 
 # Inclusion validation helpers ---------------------------------------------