+++ NEWS.md
@@ -2,6 +2,8 @@
 
 MINOR CHANGES
 
+* Improvements to documentation.
+
 * `validateIsCharacter()` is added as an alias for `validateIsString()` function.
 
 # ospsuite.utils 1.0.0