+++ NEWS.md
@@ -1,5 +1,9 @@
 # ospsuite.utils 1.0.0.9000
 
+MAJOR CHANGES
+
+* `isIncluded()` now only accepts base types as valid inputs.
+
 MINOR CHANGES
 
 * Improvements to documentation.
@@ -7,6 +11,8 @@ MINOR CHANGES
 * `validateIsCharacter()` is added as an alias for `validateIsString()` function.
 
 * `getEnumKey()` is added as an alias for `enumGetKey()` function.
+
+* Package gains `{purrr}` as a new dependency.
 
 # ospsuite.utils 1.0.0